+++ dp2_scripts/vendor_other_packages.sh
@@ -18,7 +18,10 @@ PACKAGE="libmdbodbc1_0.7~rc1-4_amd64.deb"
 THIS_FILE_DIR=$(cd $(dirname $0); pwd)
 
 cd $THIS_FILE_DIR/../vendor/
-
-wget http://us.archive.ubuntu.com/ubuntu/pool/main/m/mdbtools/$PACKAGE -O $PACKAGE
+if [ -f $THIS_FILE_DIR/../vendor/$PACKAGE ]; then
+  echo "Other packages are installed"
+else
+    wget http://us.archive.ubuntu.com/ubuntu/pool/main/m/mdbtools/$PACKAGE -O $PACKAGE
+fi
 
 echo "Done."
